+++ src/c/data.c
@@ -1662,6 +1662,18 @@ void iot_data_map_add (iot_data_t * map, iot_data_t * key, iot_data_t * val)
   }
 }
 
+void iot_data_map_add_map (iot_data_t * map, const iot_data_t * add)
+{
+  assert (map && (map->type == IOT_DATA_MAP) && add && (add->type == IOT_DATA_MAP));
+  assert (map->key_type == IOT_DATA_MULTI || (map->key_type == add->key_type));
+  iot_data_map_iter_t iter;
+  iot_data_map_iter (add, &iter);
+  while (iot_data_map_iter_next (&iter))
+  {
+    iot_data_map_add (map, iot_data_add_ref (iot_data_map_iter_key (&iter)), iot_data_add_ref (iot_data_map_iter_value (&iter)));
+  }
+}
+
 uint32_t iot_data_map_size (const iot_data_t * map)
 {
   assert (map && (map->type == IOT_DATA_MAP));